requirements.txt
Django==4.0.4
django-cors-headers==3.12.0
djangorestframework==3.13.1
Pillow==9.1.1
psycopg2==2.9.3
requests==2.27.1